agent-service/pyproject.toml
[project]
name = "partner-agent-service"
version = "0.1.0"
description = "Agent Service for partner agent integration"
authors = [
    {name = "Partner Agent Team", email = "team@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0,<0.117.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",

    "httpx>=0.25.0",
    "structlog>=23.2.0",
    # LLM backends (Phase 1: Multi-LLM support)
    "openai>=1.30.0",  # OpenAI GPT models
    "google-genai>=0.3.0",  # Google Gemini models
    # httpx already included above for Ollama support
    "partner-agent-shared-models",
    "pyyaml>=6.0.0",
    "a2a-sdk[http-server]",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=5.13.0",
    "mypy>=1.17.1",
    "types-PyYAML>=6.0.12.20250915"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_service"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::sqlalchemy.exc.MovedIn20Warning",
    "ignore::DeprecationWarning:pydantic.*:",
    "ignore:'agent_service.main' found in sys.modules:RuntimeWarning",
]

[tool.mypy]
python_version = "3.12"

# Configure mypy to use source directories for local packages
mypy_path = "../shared-models/src"

[tool.uv.sources]
partner-agent-shared-models = { path = "../shared-models" }

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=5.13.0",
    "mypy>=1.17.1",
    "types-PyYAML>=6.0.12.20250915",
]
